refactor(utils): extract getDomain + public-suffix list into utils/domain.ts
The previous commit inlined the multi-label public-suffix table and
`getDomain` helper directly into `utils/initializer.ts`, where they ran
to ~225 lines and dwarfed the rest of the file. They have nothing to do
with the page bootstrap flow that owns the rest of `initializer.ts`
(token / site / config / theme / favicon initialization), so they get
their own focused module.

Changes:

* New `src/utils/domain.ts` — owns `MULTI_LABEL_PUBLIC_SUFFIXES`,
  `isIpLike`, and `getDomain`. Same behaviour, same JSDoc.
* `src/utils/initializer.ts` — drops the inlined table + helper, imports
  `getDomain` from `./domain`, and keeps the `window.getDomain` browser-
  console alias for parity with the previous behaviour.
* `src/utils/index.ts` — re-exports `./domain` so `@/utils` consumers see
  `getDomain` through the barrel (no behavioural change for the existing
  `@/utils` imports).
* `src/components/user/Locale.vue` — switches its `getDomain` import from
  `@/utils/initializer` to the canonical `@/utils` (matches the
  pre-existing style in `Theme.vue`).

No public API change. `npx vue-tsc -b`, `eslint src`, and `npm run build`
all stay clean.
